chore: split top 4 longest running tests
