+++ packages/amplify-e2e-tests/src/__tests__/import_dynamodb_2c.test.ts
@@ -1,0 +1,152 @@
+/* eslint-disable spellcheck/spell-checker */
+/* eslint-disable import/no-extraneous-dependencies */
+
+import {
+  addAuthWithDefault,
+  AddDynamoDBSettings,
+  addDynamoDBWithGSIWithSettings,
+  amplifyPushAuth,
+  amplifyStatus,
+  createNewProjectDir,
+  deleteProject,
+  deleteProjectDir,
+  getAppId,
+  initJSProjectWithProfile,
+} from '@aws-amplify/amplify-e2e-core';
+import {
+  createDynamoDBSettings,
+  expectDynamoDBLocalAndOGMetaFilesOutputMatching,
+  expectLocalAndCloudMetaFilesMatching,
+  expectLocalAndPulledBackendConfigMatching,
+  getDynamoDBProjectDetails,
+  getShortId,
+  headlessPull,
+  importDynamoDBTable,
+} from '../import-helpers';
+
+const profileName = 'amplify-integ-test-user';
+
+describe('dynamodb import 2c', () => {
+  const projectPrefix = 'ddbimp';
+  const ogProjectPrefix = 'ogddbimp';
+
+  const projectSettings = {
+    name: projectPrefix,
+  };
+
+  const ogProjectSettings = {
+    name: ogProjectPrefix,
+  };
+
+  const dummyOGProjectSettings = {
+    name: 'dummyog1',
+  };
+
+  // OG is the CLI project that creates the dynamodb tables to import by other test projects
+  let ogProjectRoot: string;
+  let ogShortId: string;
+  let ogSettings: AddDynamoDBSettings;
+
+  // We need an extra OG project to make sure that autocomplete prompt hits in
+  let dummyOGProjectRoot: string;
+  let dummyOGShortId: string;
+  let dummyOGSettings: AddDynamoDBSettings;
+
+  let projectRoot: string;
+  let ignoreProjectDeleteErrors = false;
+
+  beforeAll(async () => {
+    ogProjectRoot = await createNewProjectDir(ogProjectSettings.name);
+    ogShortId = getShortId();
+    ogSettings = createDynamoDBSettings(ogProjectSettings.name, ogShortId);
+
+    await initJSProjectWithProfile(ogProjectRoot, ogProjectSettings);
+    await addAuthWithDefault(ogProjectRoot, {});
+    await addDynamoDBWithGSIWithSettings(ogProjectRoot, ogSettings);
+    await amplifyPushAuth(ogProjectRoot);
+
+    dummyOGProjectRoot = await createNewProjectDir(dummyOGProjectSettings.name);
+    dummyOGShortId = getShortId();
+    dummyOGSettings = createDynamoDBSettings(dummyOGProjectSettings.name, dummyOGShortId);
+
+    await initJSProjectWithProfile(dummyOGProjectRoot, dummyOGProjectSettings);
+    await addAuthWithDefault(dummyOGProjectRoot, {});
+    await addDynamoDBWithGSIWithSettings(dummyOGProjectRoot, dummyOGSettings);
+    await amplifyPushAuth(dummyOGProjectRoot);
+  });
+
+  afterAll(async () => {
+    await deleteProject(ogProjectRoot);
+    deleteProjectDir(ogProjectRoot);
+    await deleteProject(dummyOGProjectRoot);
+    deleteProjectDir(dummyOGProjectRoot);
+  });
+
+  beforeEach(async () => {
+    projectRoot = await createNewProjectDir(projectPrefix);
+    ignoreProjectDeleteErrors = false;
+  });
+
+  afterEach(async () => {
+    try {
+      await deleteProject(projectRoot);
+    } catch (error) {
+      // In some tests where project initialization fails it can lead to errors on cleanup which we
+      // can ignore if set by the test
+      if (!ignoreProjectDeleteErrors) {
+        throw error;
+      }
+    }
+    deleteProjectDir(projectRoot);
+  });
+
+  it('dynamodb headless pull successful', async () => {
+    await initJSProjectWithProfile(projectRoot, {
+      ...projectSettings,
+      disableAmplifyAppCreation: false,
+    });
+    await addAuthWithDefault(projectRoot, {});
+    await importDynamoDBTable(projectRoot, ogSettings.tableName);
+
+    await amplifyPushAuth(projectRoot);
+
+    const projectDetails = getDynamoDBProjectDetails(projectRoot);
+
+    const appId = getAppId(projectRoot);
+    expect(appId).toBeDefined();
+
+    let projectRootPull;
+
+    try {
+      projectRootPull = await createNewProjectDir('ddbimport-pull');
+
+      const envName = 'integtest';
+      const providersParam = {
+        awscloudformation: {
+          configLevel: 'project',
+          useProfile: true,
+          profileName,
+        },
+      };
+
+      const categoryConfig = {
+        storage: {
+          region: projectDetails.team.region,
+          tables: {
+            [projectDetails.storageResourceName]: projectDetails.team.tableName,
+          },
+        },
+      };
+
+      await headlessPull(projectRootPull, { envName, appId }, providersParam, categoryConfig);
+
+      await amplifyStatus(projectRoot, 'No Change');
+
+      expectLocalAndCloudMetaFilesMatching(projectRoot);
+      expectLocalAndPulledBackendConfigMatching(projectRoot, projectRootPull);
+      expectDynamoDBLocalAndOGMetaFilesOutputMatching(projectRoot, projectRootPull);
+    } finally {
+      deleteProjectDir(projectRootPull);
+    }
+  });
+});
