+++ packages/amplify-e2e-tests/src/__tests__/import_s3_2c.test.ts
@@ -1,0 +1,148 @@
+/* eslint-disable spellcheck/spell-checker */
+/* eslint-disable import/no-extraneous-dependencies */
+
+import {
+  addAuthWithDefault,
+  addS3StorageWithSettings,
+  AddStorageSettings,
+  amplifyPushAuth,
+  amplifyStatus,
+  createNewProjectDir,
+  deleteProject,
+  deleteProjectDir,
+  getAppId,
+  initJSProjectWithProfile,
+} from '@aws-amplify/amplify-e2e-core';
+import {
+  createStorageSettings,
+  expectLocalAndCloudMetaFilesMatching,
+  expectLocalAndPulledBackendConfigMatching,
+  expectS3LocalAndOGMetaFilesOutputMatching,
+  getShortId,
+  getStorageProjectDetails,
+  headlessPull,
+  importS3,
+} from '../import-helpers';
+
+const profileName = 'amplify-integ-test-user';
+
+describe('s3 import c', () => {
+  const projectPrefix = 'sssimp';
+  const ogProjectPrefix = 'ogsssimp';
+
+  const projectSettings = {
+    name: projectPrefix,
+  };
+
+  const ogProjectSettings = {
+    name: ogProjectPrefix,
+  };
+
+  const dummyOGProjectSettings = {
+    name: 'dummyog1',
+  };
+
+  // OG is the CLI project that creates the s3 bucket to import by other test projects
+  let ogProjectRoot: string;
+  let ogShortId: string;
+  let ogSettings: AddStorageSettings;
+
+  // We need an extra OG project to make sure that autocomplete prompt hits in
+  let dummyOGProjectRoot: string;
+  let dummyOGSettings: AddStorageSettings;
+
+  let projectRoot: string;
+  let ignoreProjectDeleteErrors = false;
+
+  beforeAll(async () => {
+    ogProjectRoot = await createNewProjectDir(ogProjectSettings.name);
+    ogShortId = getShortId();
+    ogSettings = createStorageSettings(ogProjectSettings.name, ogShortId);
+
+    await initJSProjectWithProfile(ogProjectRoot, ogProjectSettings);
+    await addAuthWithDefault(ogProjectRoot, {});
+    await addS3StorageWithSettings(ogProjectRoot, ogSettings);
+    await amplifyPushAuth(ogProjectRoot);
+
+    dummyOGProjectRoot = await createNewProjectDir(dummyOGProjectSettings.name);
+    dummyOGSettings = createStorageSettings(dummyOGProjectSettings.name, ogShortId);
+
+    await initJSProjectWithProfile(dummyOGProjectRoot, dummyOGProjectSettings);
+    await addAuthWithDefault(dummyOGProjectRoot, {});
+    await addS3StorageWithSettings(dummyOGProjectRoot, dummyOGSettings);
+    await amplifyPushAuth(dummyOGProjectRoot);
+  });
+
+  afterAll(async () => {
+    await deleteProject(ogProjectRoot);
+    deleteProjectDir(ogProjectRoot);
+    await deleteProject(dummyOGProjectRoot);
+    deleteProjectDir(dummyOGProjectRoot);
+  });
+
+  beforeEach(async () => {
+    projectRoot = await createNewProjectDir(projectPrefix);
+    ignoreProjectDeleteErrors = false;
+  });
+
+  afterEach(async () => {
+    try {
+      await deleteProject(projectRoot);
+    } catch (error) {
+      // In some tests where project initialization fails it can lead to errors on cleanup which we
+      // can ignore if set by the test
+      if (!ignoreProjectDeleteErrors) {
+        throw error;
+      }
+    }
+    deleteProjectDir(projectRoot);
+  });
+
+  it('storage headless pull successful', async () => {
+    await initJSProjectWithProfile(projectRoot, {
+      ...projectSettings,
+      disableAmplifyAppCreation: false,
+    });
+    await addAuthWithDefault(projectRoot, {});
+    await importS3(projectRoot, ogSettings.bucketName);
+
+    await amplifyPushAuth(projectRoot);
+
+    const projectDetails = getStorageProjectDetails(projectRoot);
+
+    const appId = getAppId(projectRoot);
+    expect(appId).toBeDefined();
+
+    let projectRootPull;
+
+    try {
+      projectRootPull = await createNewProjectDir('s3import-pull');
+
+      const envName = 'integtest';
+      const providersParam = {
+        awscloudformation: {
+          configLevel: 'project',
+          useProfile: true,
+          profileName,
+        },
+      };
+
+      const categoryConfig = {
+        storage: {
+          bucketName: projectDetails.team.bucketName,
+          region: projectDetails.team.region,
+        },
+      };
+
+      await headlessPull(projectRootPull, { envName, appId }, providersParam, categoryConfig);
+
+      await amplifyStatus(projectRoot, 'No Change');
+
+      expectLocalAndCloudMetaFilesMatching(projectRoot);
+      expectLocalAndPulledBackendConfigMatching(projectRoot, projectRootPull);
+      expectS3LocalAndOGMetaFilesOutputMatching(projectRoot, projectRootPull);
+    } finally {
+      deleteProjectDir(projectRootPull);
+    }
+  });
+});
