+++ packages/amplify-e2e-tests/src/__tests__/init_b.test.ts
@@ -1,0 +1,58 @@
+/* eslint-disable spellcheck/spell-checker */
+/* eslint-disable import/no-extraneous-dependencies */
+
+import {
+  initJSProjectWithProfile,
+  deleteProject,
+  initNewEnvWithProfile,
+  createNewProjectDir,
+  deleteProjectDir,
+  getProjectMeta,
+} from '@aws-amplify/amplify-e2e-core';
+
+describe('amplify init b', () => {
+  let projRoot: string;
+  beforeEach(async () => {
+    projRoot = await createNewProjectDir('init');
+  });
+
+  afterEach(async () => {
+    await deleteProject(projRoot);
+    deleteProjectDir(projRoot);
+  });
+
+  it('should init the project and create new env', async () => {
+    await initJSProjectWithProfile(projRoot, {});
+    const meta = getProjectMeta(projRoot).providers.awscloudformation;
+    expect(meta.Region).toBeDefined();
+    const {
+      AuthRoleName, UnauthRoleName, UnauthRoleArn, AuthRoleArn, DeploymentBucketName,
+    } = meta;
+
+    expect(UnauthRoleName).toBeIAMRoleWithArn(UnauthRoleArn);
+    expect(AuthRoleName).toBeIAMRoleWithArn(AuthRoleArn);
+    expect(DeploymentBucketName).toBeAS3Bucket(DeploymentBucketName);
+
+    // init new env
+    await initNewEnvWithProfile(projRoot, { envName: 'foo' });
+    const newEnvMeta = getProjectMeta(projRoot).providers.awscloudformation;
+
+    const {
+      AuthRoleName: newEnvAuthRoleName,
+      UnauthRoleName: newEnvUnAuthRoleName,
+      UnauthRoleArn: newEnvUnauthRoleArn,
+      AuthRoleArn: newEnvAuthRoleArn,
+      DeploymentBucketName: newEnvDeploymentBucketName,
+    } = newEnvMeta;
+
+    expect(newEnvAuthRoleName).not.toEqual(AuthRoleName);
+    expect(UnauthRoleName).not.toEqual(newEnvUnAuthRoleName);
+    expect(UnauthRoleArn).not.toEqual(newEnvUnauthRoleArn);
+    expect(AuthRoleArn).not.toEqual(newEnvAuthRoleArn);
+    expect(DeploymentBucketName).not.toEqual(newEnvDeploymentBucketName);
+
+    expect(newEnvUnAuthRoleName).toBeIAMRoleWithArn(newEnvUnauthRoleArn);
+    expect(newEnvAuthRoleName).toBeIAMRoleWithArn(newEnvAuthRoleArn);
+    expect(newEnvDeploymentBucketName).toBeAS3Bucket(DeploymentBucketName);
+  });
+});
