+++ packages/amplify-e2e-tests/src/__tests__/init_e.test.ts
@@ -1,0 +1,55 @@
+/* eslint-disable spellcheck/spell-checker */
+/* eslint-disable import/no-extraneous-dependencies */
+
+import * as fs from 'fs-extra';
+import * as path from 'path';
+import {
+  initJSProjectWithProfile,
+  deleteProject,
+  amplifyOverrideRoot,
+  amplifyPushOverride,
+  createNewProjectDir,
+  deleteProjectDir,
+  getProjectMeta,
+} from '@aws-amplify/amplify-e2e-core';
+
+import { addEnvironment } from '../environment/env';
+
+describe('amplify init e', () => {
+  let projRoot: string;
+  beforeEach(async () => {
+    projRoot = await createNewProjectDir('init');
+  });
+
+  afterEach(async () => {
+    await deleteProject(projRoot);
+    deleteProjectDir(projRoot);
+  });
+
+  it('should init the project and override root and push', async () => {
+    await initJSProjectWithProfile(projRoot, {});
+    const meta = getProjectMeta(projRoot).providers.awscloudformation;
+    expect(meta.Region).toBeDefined();
+    const {
+      AuthRoleName, UnauthRoleName, UnauthRoleArn, AuthRoleArn, DeploymentBucketName,
+    } = meta;
+
+    expect(UnauthRoleName).toBeIAMRoleWithArn(UnauthRoleArn);
+    expect(AuthRoleName).toBeIAMRoleWithArn(AuthRoleArn);
+    expect(DeploymentBucketName).toBeAS3Bucket(DeploymentBucketName);
+
+    // override new env
+    await amplifyOverrideRoot(projRoot, { testingWithLatestCodebase: true });
+    const srcOverrideFilePath = path.join(__dirname, '..', '..', 'overrides', 'override-root.ts');
+    const destOverrideFilePath = path.join(projRoot, 'amplify', 'backend', 'awscloudformation', 'override.ts');
+    fs.copyFileSync(srcOverrideFilePath, destOverrideFilePath);
+    await amplifyPushOverride(projRoot);
+    const newEnvMeta = getProjectMeta(projRoot).providers.awscloudformation;
+    expect(newEnvMeta.AuthRoleName).toContain('mockRole');
+
+    // create a new env, and the override should remain in place
+    await addEnvironment(projRoot, { envName: 'envb' });
+    const newestEnvMeta = getProjectMeta(projRoot).providers.awscloudformation;
+    expect(newestEnvMeta.AuthRoleName).toContain('mockRole');
+  });
+});
