--- packages/amplify-e2e-tests/src/__tests__/schema-auth-5.test.ts
@@ -1,83 +1,0 @@
-/* eslint-disable spellcheck/spell-checker */
-/* eslint-disable import/no-extraneous-dependencies */
-
-import {
-  initJSProjectWithProfile, deleteProject, createNewProjectDir, deleteProjectDir,
-} from '@aws-amplify/amplify-e2e-core';
-import { testSchema } from '../schema-api-directives';
-
-describe('api directives @auth batch 5.a', () => {
-  let projectDir: string;
-
-  beforeEach(async () => {
-    projectDir = await createNewProjectDir('auth5a');
-    await initJSProjectWithProfile(projectDir, {});
-  });
-
-  afterEach(async () => {
-    await deleteProject(projectDir);
-    deleteProjectDir(projectDir);
-  });
-
-  it('auth usingOidc', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'usingOidc');
-    expect(testresult).toBeTruthy();
-  });
-});
-
-describe('api directives @auth batch 5.b', () => {
-  let projectDir: string;
-
-  beforeEach(async () => {
-    projectDir = await createNewProjectDir('auth5b');
-    await initJSProjectWithProfile(projectDir, {});
-  });
-
-  afterEach(async () => {
-    await deleteProject(projectDir);
-    deleteProjectDir(projectDir);
-  });
-
-  it('auth customClaims', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'customClaims');
-    expect(testresult).toBeTruthy();
-  });
-});
-
-describe('api directives @auth batch 5.c', () => {
-  let projectDir: string;
-
-  beforeEach(async () => {
-    projectDir = await createNewProjectDir('auth5c');
-    await initJSProjectWithProfile(projectDir, {});
-  });
-
-  afterEach(async () => {
-    await deleteProject(projectDir);
-    deleteProjectDir(projectDir);
-  });
-
-  it('auth combiningAuthRules1', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'combiningAuthRules1');
-    expect(testresult).toBeTruthy();
-  });
-});
-
-describe('api directives @auth batch 5.d', () => {
-  let projectDir: string;
-
-  beforeEach(async () => {
-    projectDir = await createNewProjectDir('auth5d');
-    await initJSProjectWithProfile(projectDir, {});
-  });
-
-  afterEach(async () => {
-    await deleteProject(projectDir);
-    deleteProjectDir(projectDir);
-  });
-
-  it('auth combiningAuthRules2', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'combiningAuthRules2');
-    expect(testresult).toBeTruthy();
-  });
-});
