+++ packages/amplify-e2e-tests/src/__tests__/schema-auth-5c.test.ts
@@ -1,0 +1,26 @@
+/* eslint-disable spellcheck/spell-checker */
+/* eslint-disable import/no-extraneous-dependencies */
+
+import {
+  initJSProjectWithProfile, deleteProject, createNewProjectDir, deleteProjectDir,
+} from '@aws-amplify/amplify-e2e-core';
+import { testSchema } from '../schema-api-directives';
+
+describe('api directives @auth batch 5c', () => {
+  let projectDir: string;
+
+  beforeEach(async () => {
+    projectDir = await createNewProjectDir('auth5c');
+    await initJSProjectWithProfile(projectDir, {});
+  });
+
+  afterEach(async () => {
+    await deleteProject(projectDir);
+    deleteProjectDir(projectDir);
+  });
+
+  it('auth combiningAuthRules1', async () => {
+    const testresult = await testSchema(projectDir, 'auth', 'combiningAuthRules1');
+    expect(testresult).toBeTruthy();
+  });
+});
