+++ packages/amplify-e2e-tests/src/__tests__/schema-auth-6c.test.ts
@@ -1,0 +1,26 @@
+/* eslint-disable spellcheck/spell-checker */
+/* eslint-disable import/no-extraneous-dependencies */
+
+import {
+  initJSProjectWithProfile, deleteProject, createNewProjectDir, deleteProjectDir,
+} from '@aws-amplify/amplify-e2e-core';
+import { testSchema } from '../schema-api-directives';
+
+describe('api directives @auth batch 6c', () => {
+  let projectDir: string;
+
+  beforeEach(async () => {
+    projectDir = await createNewProjectDir('auth6');
+    await initJSProjectWithProfile(projectDir, {});
+  });
+
+  afterEach(async () => {
+    await deleteProject(projectDir);
+    deleteProjectDir(projectDir);
+  });
+
+  it('auth subscriptions2', async () => {
+    const testresult = await testSchema(projectDir, 'auth', 'subscriptions2');
+    expect(testresult).toBeTruthy();
+  });
+});
